--- src/containers/PatientDetails/HMBDiagnostic/transforms/index.ts
@@ -1,8 +1,0 @@
-import { toNumericField } from './utils';
-
-export * from './toFlowVolume';
-export * from './toPainScore';
-export { formatAuthoredDate, formatAuthoredDateTime, getChartDisplayLabel, last12 } from './utils';
-
-export const toImpact = toNumericField('impact_score');
-export const toIntensity = toNumericField('intensity');
